Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

We picked 3 because the DWARF documentation gave the impression that it was required for the new field that we use, and since DWARF version 3 is a dozen years old, surely everyone supports that by now.

I can try to come up with some test coverage for this, but not today.

Patch-set: 1
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>