net/http: Export Request.IsH2ClientPreface
Rename the existing Request.isH2Upgrade to IsH2ClientPreface (and export it). The new function name
is slightly more accurate since this is not used for upgrading from HTTP/1 to HTTP/2. It is used for
detecting when a server that only understands HTTP/1 has received a HTTP/2 connection that has
been initiated using "Starting HTTP/2 with Prior Knowledge" (RFC 7540 Section 3.4).

Change-Id: Ic495bd94db0f3373920eb80e9a260d2b55382fd3