Update patch set 1
Patch Set 1:

> It's not obvious to me that this is always correct.  This test means that an extra M was created.  At the time that we are executing this code, though, that M may be in use, and we may not have created another extra M yet (see the comment for the `needm` function in runtime/proc.go).  In that case I think we might incorrectly report a deadlock.  I'm not sure, though.

Disclaimer: I have a very superficial understanding of the Go runtime, so my analysis may be incomplete or plain wrong.

Lets first rewrite the crucial `checkdead` expression by replacing `mcount()`:

    run := sched.mnext - sched.nmfreed - sched.nmidle - sched.nmidlelocked - sched.nmsys

I note that all variables may only be updated with `sched.lock` acquired. Furthermore:

- `sched.mnext`: counts the total number of m's ever created and is only updated by `mcommoninit`, which is only called by `schedinit` and `allocm`.
- `sched.nmfreed`: counts the total number of freed m's and is only ever updated by `mexit`.
- `sched.nmidle`: counts the number of idle m's and is only ever incremented by `mput` which is called by `stopm`.
- `sched.nmidlelocked`: counts the number of locked idle m's and is only updated by `incidlelocked`. The most significant usage appears to be in `stoplockedm`.
- `sched.nmsys`: counts the number of system m's (`templateThread` and `sysmon`) and is not frequently updated.

Now for an analysis of `needm`, which is called from `cgocallback_gofunc` and `signal_unix`:

- In the `signal_unix` case the m is acquired and immediately dropped without the `sched.lock` ever acquired (and in any case it will not run on Windows).
- In the `cgocallback_gofun`c case we get a call to `cgocallbackg` and then to `cgocallbackg1`. Immediately upon caling `cgocallbackg1` we create an extra m with `newextram` / `oneNewExtraM` / `allocm`. `Allocm` will call `mcommoninit` and increment `sched.mnext`.

The danger case is any of the minus variables getting incremented between the call to `needm` / `lockextra` and `allocm` / `mcommonit`. In my analysis there appears only one point where this can happen: `exitsyscall` called from `cgocallbackg`.

`Exitsyscall` first calls `exitsyscallfast` and if that succeeds it returns immediately (without touching any of the `sched` variables of interest). Therefore `checkdead` will not fail in this case.

If `exitsyscallfast` fails, `exitsyscall` calls `exitsyscall0`, which may call `stoplockedm` or `stopm`, which will increment the `nmidlelocked` or `nmidle` variables (I believe `stoplockedm` will be called in the examined case because of the `lockOSThread` call in `cgocallbackg`). However it is important to note that the `exitsyscallfast` failure means that there are other running m's in the system, which is why we have to go through the scheduler in `exitsyscall0`. So `checkdead` will not fail in this case either.

Patch-set: 1
CC: Gerrit User 27780 <27780@62eb7196-b449-3ce5-99f1-c037f21e1705>