Update patch set 1
Patch Set 1:

Is there also going to be a Go 1.9 backport of these changes too? I guess we don't strictly need to, because by the time Go 1.11 is released, 1.11 and 1.10 will be the last two versions.

But I'm guessing there is an "upgrade period" (some period of time) immediately after a release during which you find a significant number (guessing here) on a version -2, i.e. what will be Go 1.9 at the point of the release of Go 1.11. 

The obvious case I'm thinking about here is as follows:

* Go 1.11 is released
* Package author, with major version >= 2, converts package to vgo
* Things break for any users of the package with Go 1.9
* Not wanting to break things for users during this "upgrade period", package author holds off merging vgo conversion PR until "some time later"

I don't have any concrete figures etc above; this is more just thinking aloud. My thinking is that we stand a chance of better accelerating the adoption of vgo if we also backport to 1.9

Patch-set: 1
CC: Gerrit User 16375 <16375@62eb7196-b449-3ce5-99f1-c037f21e1705>