Update patch set 1
Patch Set 1:

Nice. One idea I have for a possible additional improvement to test before we fanout a final version of changes to other map functions for go1.12 is to replace IndexByte here by a new specialised runtime.indexBucket function.

That function could get the pointer to the buckets base, and an index where to start the search.
- the function can assume there are no more than 8 byte sized tophashes and thereby only ever load 64bits starting from the bucket base then shift around the loaded hashes according to the start index
- as map functions are very hot generally in go code it could be worth to intrinsify that function to safe call overhead and allow for possible optimizations by the ssa backend across the then avoided function call boundary
- the function will likely be slightly more complex as it has to account for the start index but i think we can avoid a possible extra data cache load (vs the current CL version), avoid some of the overhead of IndexByte due to other data sizes and thereby gain some additional performance.

If this sounds reasonable enough to be accepted (if it gains better performance) we could either test the idea with this CL (to be split out later depending on the final version of the CL) if IIya has time and wants to give it a try or I would also be happy to make another prototype CL for testing so we want to further optimize this for go1.12.

Patch-set: 1