Update patch set 11
Patch Set 11:

(2 comments)

Thanks! This is great. PS 11 is much better.

> Con:
>   1. it consumes more stack space (16 bytes per frame) and is more prone to trigger stack overflow

It only needs 8 bytes for the frame pointer. Reserving 16 bytes is for alignment, right? And one word is always wasted? I think we can do better here. We can change the compiler to round the frame size from 8 mod 16 to multiple of 16. For user-written assembly function, we can simply round it up if it is 8 mod 16, but also start to accept multiple-of-16 frame size.

>   2. it tends to result in tricky bugs if users don't realize that there are extra 16 bytes for saving FP at the top of each frame.

I think we can teach the linker stack overflow check that it needs an extra word below SP on ARM64.

Patch-set: 11
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>