Update patch set 2
Patch Set 2:

Hi, thanks for working on this. I have some commentaries.

It seems to be that you have removed some of the logic out of the original test: original test was building (linking/packing?) pkg.go in two different modes, grc and grcS. Then it was compiling with -u the main package which depended on the previously built pkg.go: in the case if grcS compilation was successfull, and in the case of grc it failed with the message: cannot import unsafe package "pkg". I don't know what those modes meant at the time. 

I started working on it and got suck a bit because run.go cannot do errorcheckdir providing -u to the last pkg only, and that's what original test was doing. I might be wrong that we need to replicate original test one to one, gc has changed since 2012. Just wanted to dump my train of thoughts.

Maybe Ian can look at it and clarify things a bit.

Patch-set: 2
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 26536 <26536@62eb7196-b449-3ce5-99f1-c037f21e1705>