Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi, thanks for working on this. I have some commentaries.
> 
> It seems to be that you have removed some of the logic out of the original test: original test was building (linking/packing?) pkg.go in two different modes, grc and grcS. Then it was compiling with -u the main package which depended on the previously built pkg.go: in the case if grcS compilation was successfull, and in the case of grc it failed with the message: cannot import unsafe package "pkg". I don't know what those modes meant at the time. 
> 
> I started working on it and got suck a bit because run.go cannot do errorcheckdir providing -u to the last pkg only, and that's what original test was doing. I might be wrong that we need to replicate original test one to one, gc has changed since 2012. Just wanted to dump my train of thoughts.
> 
> Maybe Ian can look at it and clarify things a bit.

Ah, thanks!  I didn't realize the packing was (part of) the important bit here.  I thought it was just with/without the `-u` flag.  Yeah, I didn't see as much configurability in this newer method of testing (although it seems a lot nicer); I'll play around to see if it's easy to extend but other than that I'm not sure what the best course of action might be.

Thanks for taking a look and letting me know :)

Patch-set: 2
Reviewer: Gerrit User 13940 <13940@62eb7196-b449-3ce5-99f1-c037f21e1705>