Update patch set 2
Patch Set 2:

(1 comment)

> Code LGTM, though I'm confused by how the race happens. Aren't all accesses to the timer heap done under a runtime lock?

As Elias said, a user-level race condition calling Reset can cause a timer to start with one bucket and change to a different bucket in the middle of an operation.

> Perhaps the answer to the first question will answer my second question: Can't races manifest as failures other than index-out-of-bounds? For example, couldn't the heap property be invalidated?

My goal here isn't to make a racy program work.  It's only to make a racy program not crash with "panic holding locks," as that is confusing and doesn't point to the error.

Patch-set: 2