Update patch set 4
Patch Set 4: Code-Review-2

(1 comment)

Thanks for tracking this down and coming up with a fix.

That said, I don't think this is a good approach. I'm ok with a fix for this specific problem, if we can keep it light-weight. But the introduction of the isInvalid predicate suggests that we should (eventually) do this systematically, and that's where it breaks down: it will become expensive since we do these tests a lot, and a recursive test for each type is not cheap. (As an aside, it probably should be called "isValid" instead of "isInvalid").

Another problem with this approach is that a struct may have one or two fields with invalid types, but the other ones are ok. If we check first for validity of the entire struct, then we would also prevent lookup of fields that have a valid type which seems too aggressive (we've tried pretty hard for go/types to deal gracefully with errors instead of just bailing out).

A systematic approach would be nice, if we can keep it light-weight. One possibility might be to simply suppress errors that refer to invalid types (as long as there's at least one error already reported). That only costs if we have an error, not all the time.

Patch-set: 4
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2