Update patch set 4
Patch Set 4:

> Thanks for tracking this down and coming up with a fix.

Thanks to you for following up! Do CC me if you come up with a better fix, as I'll be interested to know :)

I'm not familiar enough with go/types to easily come up with a better approach, and I'm a bit short on time at the moment, so I'll abandon this and leave this to someone else for now.

Patch-set: 4
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>