Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Thanks, Robert - this indeed seems like a cheaper and simpler solution. I wonder if the heuristic will ever cause issues, but I guess we'll see :)
> 
> Daniel, in the issue, you said "cmd/compile does not have this issue". Do you know how it avoids it? Does it use a similar strategy as this fix?

Both cmd/compile and go/types use a comparable recursive type-checking algorithm but the actual implementation is vastly different and the details are not comparable, resulting in different error reporting. Basically, it's relatively easy to detect an error, but it's rather hard to continue in a meaningful way in the presence of an error. go/types tries harder than cmd/compile (after all, when we compile we could just give up as soon as we see an error; but go/types is used in situations where we want to type-check partially correct code). Follow-on errors are errors that occur because of a prior error having ripple-effects. Sometimes they can be masked by pretending the prior error didn't happen, sometimes we filter them (like here). The goal is to present the user with the important errors; it doesn't matter too much if some errors are dropped. For instance, the compiler doesn't report more than one error per line (it just throws them away), but go/types does.

Patch-set: 3