Update patch set 2
Patch Set 2:

While I have no objections to fixing things in the short term, now that there have been multiple operating-systems specific tweaks to get the test to work, it should be rewritten to be portable. Surely that's not difficult.

Since we're only looking for equality, modulo spaces, it's an easy job to write in line. Faster to run, too. And totally portable. @dsymonds?

Patch-set: 2
CC: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>