internal/poll: specify current file position when calling TransmitFile
Current SendFile implementation assumes that TransmitFile starts from
the current file position. But that appears not true for some Windows
version.

TransmitFile documentation

https://msdn.microsoft.com/en-us/library/windows/desktop/ms740565(v=vs.85).aspx

suggests

```
You can use the lpOverlapped parameter to specify a 64-bit offset within
the file at which to start the file data transfer by setting the Offset
and OffsetHigh member of the OVERLAPPED structure.
```

so do as it advises.

Fixes #25722

Change-Id: I241d3bf76d0d5590d4df27c6f922d637068232fb