Update patch set 2
Patch Set 2:

I discussed these XML changes with Russ and others and it sounds like we want to prioritize these (and JSON) for after Go 1.11 and after vgo. We just didn't have the review bandwidth in this area this cycle and can't accept the risk of regressions this late in Go 1.11. I'm going to mark this for Go 1.12 for now. Sorry.

Patch-set: 2
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>