Update patch set 2
Patch Set 2: Code-Review+2

Is the gist here something like:

We want our segments' VAs to be aligned by something the kernel will like for the start of an mmap, which is platform dependent and controlled by FlagRound. There's no particular reason the layout in the file should be aligned, but nonetheless it had an assumption that segments were aligned *relative to each other* in the file. That is, the distance between the start of any two segments will be a multiple of FlagRound, but there will be some prefix in the file for the ELF header or whatever. The previous version of the CL aligned the segments in the file, rather than relative to each other, breaking that assumption. When a later part of the linker "fixed" the alignments, it threw everything out of whack and broke.

?

Patch-set: 2
Reviewer: Gerrit User 17090 <17090@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2