html/template: ignore single untyped nil values
Previous versions of Go like 1.10 did not support using untyped nils as
values in templates very well. For example, using one in a pipeline
resulted in no value being passed, as it was identified as a lack of
arguments.

Since then, we have fixed text/template to fully support untyped nils.

An unforeseen consequence, however, is that html/template now outputs an
escaped "<nil>" when printing an untyped nil, when it used to output
nothing. This is because the template used to skip that value, and it no
longer does in 1.11.

To avoid the change in behavior in 1.11 in the HTML code, make it always
skip a single untyped nil value, just like 1.10. The same is done in
text/template.HTMLEscaper, to keep behavior consistent.

An old and fairly niche html/template test needed updating, as it was
locking in behavior that an untyped nil field should print an escaped
"<nil>". The original fix was for a crash, and html/template has always
ignored untyped nils in all other cases, so it seems safe to bring this
together.

Fixes #25875.

DO NOT SUBMIT - we haven't yet decided if this is the right path
forward.

Change-Id: I34cde7e215ec821e7393300c4742f420c988b511