Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> Is the gist here something like:
> 
> We want our segments' VAs to be aligned by something the kernel will like for the start of an mmap, which is platform dependent and controlled by FlagRound. There's no particular reason the layout in the file should be aligned, but nonetheless it had an assumption that segments were aligned *relative to each other* in the file. That is, the distance between the start of any two segments will be a multiple of FlagRound, but there will be some prefix in the file for the ELF header or whatever. The previous version of the CL aligned the segments in the file, rather than relative to each other, breaking that assumption. When a later part of the linker "fixed" the alignments, it threw everything out of whack and broke.
> 
> ?

More or less. The condition that's now checked in layout is what really matters: to make mmap'ing segments work, the VA and file offset must be congruent mod the page size. We don't really know the page size, but *FlagRound is a good enough substitute.

Segment VAs are aligned to such a large boundary because they may need to be mapped with incompatible permissions. Fancier linkers understand that segments can be packed right next to each other in the file as long as their VAs are far enough apart to avoid the need to map one page with different permissions. In this case, a page of the file can be mapped multiple times (once for segment N containing the unused head of segment N+1 and once for segment N+1 containing the unused tail of segment N). Since our linker doesn't do these VA gap tricks, it has to align the segment VAs instead.

Patch-set: 2
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>