cmd/compile: display Go code for a function in ssa.html
This CL adds the "sources" column at the beginning of SSA table.
This column displays the source code for the function being passed
in the GOSSAFUNC env variable.

Also UI was extended so that clicking on particular line will
highlight all places this line is referenced.

JS code was cleaned a bit.

This CL does not handle inlined functions. That is TBD.

Change-Id: Ic7833a0b05e38795f4cf090f3dc82abf62d97026