cmd/go: better quote the cfg.GOROOT output
See what happened to me (in a Makefile)

    _GOROOT:=${GOROOT}       # save the original value
    GOROOT:=$(shell pwd)/goroot
    -include local_config.mk
    [...]

and in local_config.mk

    GOROOT:=${_GOROOT}

The problem now is, that because of the *wrong* "inline"
comment, _GOROOT gets trailing spaces and a `go env` test
gives (in the make context)

    go: cannot find GOROOT directory: /home/ingo/src/go.src

where the trailing spaces are invisible and the answer
looks just wrong.

With this patch the result says

    go: cannot find GOROOT directory: '/home/ingo/src/go.src      '

which is readable and correct.

Change-Id: Ie204c97886266494e3e35777a416ea63adce47f9