image/gif: fix acceptance of one non-significant byte in data
LZW allows a 00 byte to occur at the end of the block.
Image data reader close method was handling this byte
as comments detail. This byte is not the clear code.

The decoder holds the validity of the possible extra byte
as image data processing occurs after reading its description.

The rule on frame and image bounds now allows a one byte difference
which validity is checked later.

One related test added.

Fixes #20856

Change-Id: I41378f924f61f742ff9f3fa37604e7edd7b11d5f