image/gif: fix acceptance of one non-significant byte
LZW allows a 00 byte to occur at the end of the block. Close method
is handling the case and comments detail the rational.
This byte is not the clear code.

When the byte occurs, the size of the image might be too large by
1 byte. Calculations of image bounds are adapted.

If the byte has a value other than 00, an error will be returned.
Existing tests required one change for that purpose.

Fixes #20856

Change-Id: I41378f924f61f742ff9f3fa37604e7edd7b11d5f