image/gif: fix acceptance of one non-significant byte
LZW allows a 00 byte to occur at the end of the block. Close method
was handling this case as comments detail.
This byte is not the clear code.

The decoder holds if this extra byte is legit or not as image
data processing occurs later.

The rule on frame and image bounds allows a one byte difference
which is checked after processing of image data.

One test added.

Fixes #20856

Change-Id: I41378f924f61f742ff9f3fa37604e7edd7b11d5f