Update patch set 4
Patch Set 4:

> > Once crypto/x509 grows support for PSS pubkeys, then the crypto/tls layer could be extended further.
> 
> Ah I see. Then I'd like a test making sure certificates with PSS keys fail entirely. Otherwise when crypto/x509 support is added we'll suddenly accept them with rsa_pss_rsae_xxx codepoints.

It is not straightforward to test that for now. The pickSignatureAlgorithm function currently takes a `crypto.PublicKey` while the extra PSS parameters are part of the certificate. Both RSASSA-PSS and idEncryption public keys use the same RSAPublicKey definition from https://tools.ietf.org/html/rfc8017#appendix-A.1.1

Such a test will depend on how certificates with RSASSA-PSS public keys are serialized. One way to implement the data structure for PSS public keys is to extend rsa.PublicKey with a new "Parameters" (or "PSSParameters") field. This is close to how dsa.PublicKey is implemented.

However, it risks that the public key is used for wrong purposes (crypto/tls could treat it as a cert with idEncryption pubkey). An alternative is to create a new rsa.PSSPublicKey. This will make crypto/tls reject the certificate (as desired).

Do you still think that some kind of test makes sense for now? As an alternative, this patch can be postponed/dropped until crypto/x509 gains full support, but at the moment I don't have that much time to work on that.

Patch-set: 4
Reviewer: Gerrit User 24681 <24681@62eb7196-b449-3ce5-99f1-c037f21e1705>