Update patch set 1
Patch Set 1:

We may want to go ahead and do this for Go 1.11. It's obviously a more complicated change than the parent CL, but I'm also not convinced I got all of the code paths in the previous CL since we set up stacks in a surprising variety of ways. This CL is a much cleaner and more complete solution.

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>